chore(style): remove old style casts and cross sign comparisons
Enable warnings for both. Favour casting to signed rather than casting to
unsigend for comparisons. Per isocpp's core guidelines ES.102, signed
arithmetic gives more expected results. If we need one extra bit of range,
using longer signed types achives that.

Fix #6010
Fix #6012

+++ src/model/sessionchatlog.cpp
@@ -158,12 +158,12 @@ SearchResult SessionChatLog::searchForward(SearchPos startPos, const QString& ph
 
         auto numMatches = 0;
         QRegularExpressionMatch lastMatch;
-        while (match.isValid() && numMatches <= currentPos.numMatches && match.hasNext()) {
+        while (match.isValid() && numMatches <= static_cast<int>(currentPos.numMatches) && match.hasNext()) {
             lastMatch = match.next();
             numMatches++;
         }
 
-        if (numMatches > currentPos.numMatches) {
+        if (numMatches > static_cast<int>(currentPos.numMatches)) {
             SearchResult res;
             res.found = true;
             res.pos.logIdx = key;
@@ -223,13 +223,13 @@ SearchResult SessionChatLog::searchBackward(SearchPos startPos, const QString& p
         while (match.isValid() && match.hasNext()) {
             auto currentMatch = match.next();
             totalMatches++;
-            if (currentPos.numMatches == 0 || currentPos.numMatches > numMatchesBeforePos) {
+            if (currentPos.numMatches == 0 || static_cast<int>(currentPos.numMatches) > numMatchesBeforePos) {
                 lastMatch = currentMatch;
                 numMatchesBeforePos++;
             }
         }
 
-        if ((numMatchesBeforePos < currentPos.numMatches || currentPos.numMatches == 0)
+        if ((numMatchesBeforePos < static_cast<int>(currentPos.numMatches) || currentPos.numMatches == 0)
             && numMatchesBeforePos > 0) {
             SearchResult res;
             res.found = true;
@@ -285,7 +285,7 @@ std::vector<IChatLog::DateChatLogIdxPair> SessionChatLog::getDateIdxs(const QDat
         ret.push_back(std::move(pair));
 
         dateIt = dateIt.addDays(1);
-        if (startDate.daysTo(dateIt) > maxDates && maxDates != 0) {
+        if (startDate.daysTo(dateIt) > static_cast<long>(maxDates) && maxDates != 0) {
             break;
         }
     }
